pyproject.toml
[project]
name = "testpydanticai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai-slim[mcp,openai]>=1.35.0",
    "restate-sdk[serde]>=0.14.2",
    "uvicorn>=0.38.0",
]

